week05-homework/pyproject.toml
[project]
name = "week05-homework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"active>=0.0.5a0",
	"fastmcp>=2.14.5",
	"langchain-mcp-adapters>=0.1.0",
	"langchain>=1.2.8",
	"langchain-core>=1.2.9",
	"langchain-deepseek>=1.0.1",
	"langgraph>=1.0.7",
	"langgraph-checkpoint-sqlite>=2.0.0",
	"python-dotenv",
	"tavily>=1.1.0",
	"uvicorn>=0.30.0",
]
